use ActionController::Base.authenticate to require autentication for your actions.
the authenticate macro (class method) when called alone require a valid session for any route except the new session and create session. To modify this behaviour use :only and :except in the same manner you would for filters. For example, :except => [:index] will allow the index action of the controller to be invoked without a valid session. Be careful with this however, right now for example, if you add this to your application controller any index action will will be able to invoked. There will be better ways for doing this soon.

Note: authenticate cannot be called on ActionController::Base it must be called on ApplicationController or its Subclasses.