skip_authenticate! allows actions to be called without a valid_session
skip_authenticate! macro takes the :only and :except arguments to specify exactly which actions in a controller can be run w/o a valid session. If not passed any arguments skip_authenticate! will allow any action for that controller to be run w/o authentication.

I suggest you add the authenticate macro to your ApplicationController and then use skip_authenticate! when needed in subclasses of ApplicationController. If you have an action, application wide, that should be allowed to run w/o authentication then simply use the :only and :except options that can be passed to authenticate.