remember me when I login
you will need to manually add the columns added to the migration in the previous commit to your model or rollback the accounts table migration, regenerate it and migrate up again. The new columns are remember_token and remember_expiry. In the future there will be a way to opt out of this.